tasks 5 and 6 new
