+++ src/AddImportMastersAction.java
@@ -1,0 +1,2 @@
+package PACKAGE_NAME;public class AddImportMasters {
+}
